+++ backend/src/baserow/contrib/database/fields/field_types.py
@@ -85,6 +85,41 @@ class NumberFieldType(FieldType):
                 positive=not instance.number_negative
             )
 
+    def get_alter_column_type_function(self, connection, instance):
+        if connection.vendor == 'postgresql':
+            decimal_places = 0
+            if instance.number_type == NUMBER_TYPE_DECIMAL:
+                decimal_places = instance.number_decimal_places
+
+            function = f"round(p_in::numeric, {decimal_places})"
+
+            if not instance.number_negative:
+                function = f"greatest({function}, 0)"
+
+            return function
+
+        return super().get_alter_column_type_function(connection, instance)
+
+    def after_update(self, field, old_field, model, old_model, connection,
+                     altered_column):
+        """
+        The allowing of negative values isn't stored in the database field type. If
+        the type hasn't changed, but the allowing of negative values has it means that
+        the column data hasn't been converted to positive values yet. We need to do
+        this here. All the negatives values are set to 0.
+        """
+
+        if (
+            not altered_column
+            and not field.number_negative
+            and old_field.number_negative
+        ):
+            model.objects.filter(**{
+                f'field_{field.id}__lt': 0
+            }).update(**{
+                f'field_{field.id}': 0
+            })
+
 
 class BooleanFieldType(FieldType):
     type = 'boolean'