+++ changelog.md
@@ -10,3 +10,6 @@
 * The next field is now selected when the tab character is pressed when a field is
   selected.
 * Changed the styling of the notification alerts.
+* Fixed error when changing field type and the data value wasn't in the correct
+  format.
+* Update the field's data values when the type changes.